requirements.txt
playwright==1.42.0
pytest==7.4.0
pytest-playwright==0.4.0
allure-pytest==2.13.2
requests==2.31.0